refactor rest_app
