--- dj_rest/README.txt
@@ -1,1 +1,0 @@
-homework project for django rest framwork
