+++ dj_rest/config/urls.py
@@ -4,8 +4,6 @@ from django.urls import path, include
 from django.views.generic import RedirectView
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('todoapp.urls', namespace='todoapp')),
+    path('api/', include('usersapp.urls', namespace='usersapp')),
     path("", RedirectView.as_view(url="api/")),
 ]
-
-
