+++ dj_rest/todoapp/admin.py
@@ -1,9 +1,3 @@
- 
 from django.contrib import admin
-# from todo.models import User_models
 
-
-# @admin.register(User_models.User)
-# class UserAdmin(admin.ModelAdmin):
-#     list_display = ['username', 'first_name', 'last_name', 'email']
-#     ordering = ['username']
+# Register your models here.
