+++ dj_rest/todoapp/views.py
@@ -1,7 +1,3 @@
-from rest_framework.viewsets import ModelViewSet
-from todoapp.models import UserModel
-from todoapp.serializers import AuthorModelSerializer
+from django.shortcuts import render
 
-class AuthorModelViewSet(ModelViewSet):
-    queryset = UserModel.objects.all()
-    serializer_class = AuthorModelSerializer
+# Create your views here.
