+++ dj_rest/usersapp/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+# from todo.models import User_models
+
+
+# @admin.register(User_models.User)
+# class UserAdmin(admin.ModelAdmin):
+#     list_display = ['username', 'first_name', 'last_name', 'email']
+#     ordering = ['username']
