+++ dj_rest/usersapp/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class TodoappConfig(AppConfig):
+class UsersappConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'todoapp'
+    name = 'usersapp'
